add exceptions

+++ src/main/java/com/uNow/controller/ActivityController.java
@@ -3,6 +3,8 @@ package com.uNow.controller;
 
 import com.uNow.entities.Activity;
 import com.uNow.entities.User;
+import com.uNow.exceptions.ActivityNotFoundException;
+import com.uNow.exceptions.BadRequestException;
 import com.uNow.exceptions.UserNotFoundException;
 import com.uNow.repositories.ActivityRepository;
 import com.uNow.repositories.UserRepository;
@@ -36,6 +38,14 @@ public class ActivityController {
     }
 
     @CrossOrigin
+    @GetMapping("/getActivity/{activityId}")
+    public Activity findById(@PathVariable("activityId") long activityId) throws ActivityNotFoundException {
+        if (activityRepository.findById(activityId).get() == null)
+            throw new ActivityNotFoundException();
+        return activityRepository.findById(activityId).get();
+    }
+
+    @CrossOrigin
     @GetMapping("/{userId}")
     public List<Activity> findAllByUser(@PathVariable("userId") Long userId )  throws UserNotFoundException {
         User user = userRepository.findById(userId).get();
@@ -44,11 +54,24 @@ public class ActivityController {
         return activityRepository.findByUser(user);
     }
 
+
+    @CrossOrigin
+    @PutMapping
+    public void addLike(@RequestBody Activity activity) throws BadRequestException {
+        if (!activityRepository.existsById(activity.getId()))
+            throw new BadRequestException();
+        activity.setLikes(activity.getLikes() + 1);
+        activityRepository.save(activity);
+    }
+
+
     @CrossOrigin
     @DeleteMapping("/{activityId}")
-    public void deleteActivity(@PathVariable("activityId") long activityId) {
+    public void deleteActivity(@PathVariable("activityId") long activityId) throws ActivityNotFoundException {
         Activity activityToDelete = activityRepository.findById(activityId).get();
-        activityRepository.delete(activityToDelete);
+        if (activityToDelete != null)
+            activityRepository.delete(activityToDelete);
+        throw new ActivityNotFoundException();
     }
 
     @ExceptionHandler
@@ -56,17 +79,14 @@ public class ActivityController {
     public void userNotFoundHandler(UserNotFoundException e) {
     }
 
-    @CrossOrigin
-    @GetMapping("/getActivity/{activityId}")
-    public Activity getActivityById(@PathVariable("activityId") long activityId) {
-        return activityRepository.findById(activityId).get();
+    @ExceptionHandler
+    @ResponseStatus(HttpStatus.NOT_FOUND)
+    public void activityNotFoundHandler(ActivityNotFoundException e) {
     }
 
-    @CrossOrigin
-    @PutMapping
-    public void addLike(@RequestBody Activity activity) {
-        activity.setLikes(activity.getLikes() + 1);
-        activityRepository.save(activity);
+    @ExceptionHandler
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    public void badRequestHandler(ActivityNotFoundException e) {
     }
 
 }
